+++ ImaGE/source/Core/TempObject.h
@@ -7,7 +7,7 @@
 
 struct Object
 {
-  glm::mat4 mdlTransform;
+  glm::dmat4 mdlTransform;
   Component::Transform transform;
   glm::vec4 clr;
   
@@ -15,16 +15,16 @@ struct Object
   bool modified, collided;
 
   Object(Graphics::MeshType type, glm::vec4 const& color = { 1.f, 0.5f, 0.2f, 1.f },
-    glm::vec3 const& _pos = {}, glm::vec3 const& _scale = { 1.f, 1.f, 1.f });
+    glm::dvec3 const& _pos = {}, glm::dvec3 const& _scale = { 1.0, 1.0, 1.0 });
 
-  Object(std::string const& filePath, glm::vec3 const& _pos = {},
-    glm::vec3 const& _scale = { 1.f, 1.f, 1.f });
+  Object(std::string const& filePath, glm::dvec3 const& _pos = {},
+    glm::dvec3 const& _scale = { 1.0, 1.0, 1.0 });
 
   void Update(float deltaTime);
 
-  inline glm::vec3& GetPos() noexcept { return transform.worldPos; }
-  inline glm::vec3 const& GetPos() const noexcept { return transform.worldPos; }
-  inline glm::vec3& GetScale() noexcept { return transform.worldScale; }
-  inline glm::vec3 const& GetScale() const noexcept { return transform.worldScale; }
+  inline glm::dvec3& GetPos() noexcept { return transform.worldPos; }
+  inline glm::dvec3 const& GetPos() const noexcept { return transform.worldPos; }
+  inline glm::dvec3& GetScale() noexcept { return transform.worldScale; }
+  inline glm::dvec3 const& GetScale() const noexcept { return transform.worldScale; }
 };
